Fix negative mouse-wheel movements, swap actions for wheelup/wheeldown
Now wheeldown should swap weapon, while wheelup will swap gadget

+++ src/xenia/ui/window_win.cc
@@ -524,7 +524,7 @@ LRESULT Win32Window::WndProc(HWND hWnd, UINT message, WPARAM wParam,
 
         auto e = MouseEvent(this, MouseEvent::Button::kNone, mouseData.lLastX,
                             mouseData.lLastY, mouseData.usButtonFlags,
-                            mouseData.usButtonData);
+                            (int16_t)mouseData.usButtonData);
         OnRawMouseMove(&e);
       }
